@@ -20,25 +20,22 @@
     "irc": "irc://irc.freenode.net/imbo"
   },
   "require": {
-    "php": ">=7.2.5",
+    "php": "^7.4",
     "ext-json": "*",
-    "behat/behat": "^3.5",
-    "guzzlehttp/guzzle": "^6.5",
-    "beberlei/assert": "^3.2",
-    "firebase/php-jwt": "^5.0"
+    "beberlei/assert": "^3.3",
+    "behat/behat": "^3.8",
+    "firebase/php-jwt": "^5.2",
+    "guzzlehttp/guzzle": "^7.3"
   },
   "require-dev": {
-    "symfony/process": "^5.0",
-    "phpunit/phpunit": "^8.5",
-    "slim/slim": "^4.4",
-    "slim/psr7": "^1.0",
-    "tuupola/slim-basic-auth": "^3.2",
-    "phpstan/phpstan": "^0.12",
-    "phpstan/phpstan-phpunit": "^0.12",
-    "phpstan/extension-installer": "^1.0",
-    "vimeo/psalm": "^3.12",
-    "psalm/plugin-phpunit": "^0.12",
-    "alexeyshockov/guzzle-psalm-plugin": "^0.2.1"
+    "alexeyshockov/guzzle-psalm-plugin": "^0.3.1",
+    "phpunit/phpunit": "^9.5",
+    "psalm/plugin-phpunit": "^0.15.1",
+    "slim/psr7": "^1.3",
+    "slim/slim": "^4.7",
+    "symfony/process": "^5.2",
+    "tuupola/slim-basic-auth": "^3.3",
+    "vimeo/psalm": "^4.7"
   },
   "autoload": {
     "psr-4": {
@@ -51,27 +48,27 @@
     }
   },
   "scripts": {
-    "clean": "rm -rf build",
     "lint": "for file in `git ls-files '*php'`; do php -l $file; done",
-    "test:phpunit": "vendor/bin/phpunit",
-    "test:phpunit:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
-    "test:behat": "vendor/bin/behat --strict",
+    "phpunit": "vendor/bin/phpunit",
+    "phpunit:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
+    "behat": "vendor/bin/behat --strict",
+    "psalm": "vendor/bin/psalm",
     "test": [
-      "@test:phpunit",
-      "@test:behat"
+      "@phpunit",
+      "@behat"
     ],
-    "sa:phpstan": "vendor/bin/phpstan analyse",
-    "sa:psalm": "vendor/bin/psalm",
     "sa": [
-      "@sa:phpstan",
-      "@sa:psalm"
+      "@psalm"
     ],
     "ci": [
       "@lint",
-      "@sa",
-      "@test"
+      "@test",
+      "@sa"
     ],
     "dev": "php -S localhost:8080 -t ./features/bootstrap > server.log 2>&1",
     "docs": "cd docs; make html"
+  },
+  "config": {
+    "sort-packages": true
   }
 }
